egl-ext: Add capability to map window front buffer
Added a function for mapping and unmapping the front buffer of a window
for output verification purposes. This mechanism replaces the earlier
way of mapping the system frame buffer, because the previous mechanism
was found to be unreliable in an automated testing environment where
other applications may also interact with the system frame buffer.

This patch set implements front buffer mapping with XGetImage(). Only
read-only mapping is currently supported.

Signed-off-by: Sami Kyöstilä <sami.kyostila@nokia.com>