egl-ext: Simplifies an internal API for solid quad drawing.
Simplifies an internal API which is used for solid quad drawing.
The color channel values are now given as GLfloats.

Also corrects the order of color channels and alpha in the vertex
array for solid quad rendering.

Signed-off-by: Tero Tiittanen <tero.tiittanen@nokia.com>

Changed alpha channel from 0.0f to 1.0f to avoid possible problems with
blending in the future.

Signed-off-by: Sami Kyöstilä <sami.kyostila@nokia.com>